🅿️[P] 2 apps to allow previewing of custom, text-based file extensions
This is a feature to address Issue #87: "Add support for dynamic and unknown UTI types": https://github.com/whomwah/qlstephen/issues/87
In macOS 10.14(?) and later, QLStephen doesn't appear to work for text-based files that have a filename extension that is unclaimed by any app or bundle present on their system. For example, if no application or bundle on the system claims the “.leaf” filename extension, and assuming they are text files, QLStephen won’t be used to render the text.

One way to solve this might be to include a default list of filename extensions that we’d define as text. However, this could be problematic as one person’s `.map` file could be text while another’s could be binary. The default list could also conflict with other apps on the user’s system who do define these filename extensions differently.

The solution I came up with is based on the assumption that only the end-user is going to know what filename extensions they come across and whether those files are text-based or not. It uses 2 different apps:

1. `QLStephen.app`: the primary app you interact with. It holds a list of the filename extensions that you've mapped to being treated as text. To map a particular filename extension to be treated as text, we declare that it conforms to “public.plain-text”. By doing so, we then allow the system’s built-in text quicklook plugin to be used to preview the contents. To implement these file mappings on the system, QLStephen.app uses a second application, “QLStephen Dummy.app”, in which the mappings are declared in the `UTExportedTypeDeclarations` of its Info.plist file. QLStephen.app also adds a Service so that you can right-click/control-click on any file in the Finder, and choose “Quick Look with QLStephen ✅” (yes, this is a bit misleading, I guess). This automatically opens it in QLStephen.app, and if it’s found to be text-based, is added to the list of file extension mappings. Then, after updating the dummy app, you can deselect and then reselect the file in the Finder to preview its contents as text.

2. `QLStephen Dummy.app`:  This app is created in the ~/Applications/ folder. As mentioned previously, this is a dummy app that declares the file extension mappings in its ‘UTExportedTypeDeclarations` key in its Info.plist file. After modifying the Info.plist of the installed app, it is codesigned to run locally.

Some notes/known issues:
• Despite having a “Kind” column for you to fill out a file extension’s description, it appears the Finder will always refer to the kind as “Plain Text Document”.
• There are some filename extensions that still won’t preview even though they are text-based: `.css` files and `.strings` files are 2 examples I’ve encountered.